Move stats-related plots routines from aimodel/ to statutil/. Do *not* move the aimodel-plots
