+++ pdr_backend/statutil/autocorrelation_plotter.py
@@ -4,7 +4,7 @@ from enforce_typing import enforce_types
 from plotly.subplots import make_subplots
 from statsmodels.tsa.stattools import adfuller
 
-from pdr_backend.aimodel.autocorrelation import (
+from pdr_backend.statutil.autocorrelation import (
     AutocorrelationPlotdata,
     _add_corr_traces,
 )
